task2_user_admin_ai_feedback_system/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
httpx>=0.27.0
python-dotenv>=1.0.1
supabase>=2.6.0